training/requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
redis>=5.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
asyncio>=3.4.3
lightgbm>=4.5.0
